dfguard/ml-worker/requirements.txt
torch==2.1.0
torchvision==0.16.0
facenet-pytorch==2.5.3
Pillow==10.0.1
numpy==1.24.4
boto3==1.28.57
pymongo==4.5.0
dnspython==2.4.2
requests==2.31.0
tqdm==4.66.1
